changing a build warning

+++ engine-v2/src/main.cpp
@@ -63,7 +63,7 @@ int main(void) {
 
 	// Load a blue texture to act as the movement range indicator.
 	{
-		int index = gs.textures.size();
+		size_t index = gs.textures.size();
 		gs.texture_handles.insert({ "Movement", index });
 		gs.textures.push_back(LoadTextureFromImage(GenImageColor(gs.entity_scale, gs.entity_scale, SKYBLUE)));
 	}
